requirements.txt
streamlit
Pillow
opencv-python
numpy
ultralytics
torch
basicsr
facexlib
